Hot Fix 1
I had realized as soon as I stepped away from my computer that I don't need to modify the changes made to v2, v3, & vAlpha within the update() function. This hot fix corrects them, as we should only be concerned with how we scroll the Hue from clr1 to clr2.